requirements.txt
# Streamlit App dependencies
streamlit
numpy
pandas
scikit-learn
dlib-bin
git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0
supabase
bcrypt
segno
pillow
librosa
resemblyzer

# Flask Landing Page dependencies
flask
gunicorn

# Utilities
python-dotenv
